+++ sway-core/src/type_system/mod.rs
@@ -1,34 +1,34 @@
+mod binding;
 mod collect_types_metadata;
 mod copy_types;
 mod create_type_id;
+mod engine;
+mod id;
+mod info;
 mod length;
+mod mapping;
 mod occurs_check;
 mod replace_self_type;
 mod trait_constraint;
 mod type_argument;
-mod type_binding;
-mod type_engine;
-mod type_id;
-mod type_info;
-mod type_mapping;
 mod type_parameter;
 mod unconstrained_type_parameters;
-mod unify;
+mod unifier;
 mod unify_check;
 
+pub(crate) use binding::*;
 pub(crate) use collect_types_metadata::*;
 pub(crate) use copy_types::*;
 pub(crate) use create_type_id::*;
+pub use engine::*;
+pub use id::*;
+pub use info::*;
 pub use length::*;
+pub(crate) use mapping::*;
 use occurs_check::*;
 pub(crate) use replace_self_type::*;
 pub(crate) use trait_constraint::*;
 pub use type_argument::*;
-pub(crate) use type_binding::*;
-pub use type_engine::*;
-pub use type_id::*;
-pub use type_info::*;
-pub(crate) use type_mapping::*;
 pub use type_parameter::*;
 pub(crate) use unconstrained_type_parameters::*;
